+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## Developer Changelog
+## Developer Changelog 
+
+### May 17, 2022 DESCW-382
+* frontend
+  * implemented react-query for project details api
+
 
 ### May 16, 2022 DESCW-362
 * frontend